model_server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
transformers==4.37.0
torch==2.1.2
accelerate==0.26.1
bitsandbytes==0.42.0
pillow==10.2.0
pydantic==2.5.3
python-multipart==0.0.6
sentencepiece==0.1.99
protobuf==4.25.2
